+++ example/main.dart
@@ -4,6 +4,26 @@ void main() async {
   ///use Google Translate
   final gt = SimplyTranslator(EngineType.google);
 
+  ///if you do not specify the source language it is automatically selecting it depending on the text
+  ///if you do not specify the target language it is automatically English
+
+  ///get "hello" as an Audio-Url
+  ///uses always Google TTS as Libretranslate doesnt support TTS, gives same result
+  ///you can use https://pub.dev/packages/audioplayers to play it
+  print(gt.getTTSUrl("hello", "en"));
+  //https://simplytranslate.org/api/tts/?engine=google&lang=en&text=hello
+
+  ///using Googletranslate:
+  ///short form to only get translated text as String, also shorter code:
+  String textResult = await gt.trSimply("Er läuft schnell.", "de", 'en');
+
+  print(textResult);
+  //He walks fast.
+
+  ///get time it took to translate
+  print(await gt.speedTest(gt.trSimply, "Er läuft schnell."));
+  //1278ms
+
   ///use Libretranslate
   final lt = SimplyTranslator(EngineType.libre);
 
@@ -12,38 +32,28 @@ void main() async {
   //[simplytranslate.org, st.tokhmi.xyz, translate.josias.dev, ...
 
   ///update instances with the API
-  gt.updateInstances();
+  gt.updateSimplyInstances();
 
   ///check if instance is working
-  print(await gt.isInstanceWorking("simplytranslate.pussthecat.org"));
+  print(await gt.isSimplyInstanceWorking("simplytranslate.pussthecat.org"));
   //true
 
   /// find other instances under https://simple-web.org/projects/simplytranslate.html
   ///change instance (defaut is simplytranslate.org)
   gt.setInstance = "simplytranslate.pussthecat.org";
 
-  ///if you do not specify the source language it is automatically selecting it depending on the text
-  ///if you do not specify the target language it is automatically English
-
-  ///get "hello" as an Audio-Url
-  ///uses always Google TTS as Libretranslate doesnt support TTS, gives same result
-  ///you can use https://pub.dev/packages/audioplayers to play it
-  print(gt.getTTSUrl("hello", "en"));
-  print(lt.getTTSUrl("hello", "en"));
-  //https://simplytranslate.org/api/tts/?engine=google&lang=en&text=hello
-
   ///using Libretranslate
   ///only text translation avaliable
   ///short form to only get translated text as String, also shorter code:
-  String textResult = await lt.tr("Er läuft schnell.", "de", 'en');
+  textResult = await lt.trSimply("Er läuft schnell.", "de", 'en');
 
   ///is the same as
-  textResult = await lt.tr("Er läuft schnell.");
+  textResult = await lt.trSimply("Er läuft schnell.");
   print(textResult);
   //he's running fast.
 
   ///long form, switching to next instance, 4 retries if it fails (default 1)
-  var libTransl = await lt.translate(
+  var libTransl = await lt.translateSimply(
       "The dispositions were very complicated and difficult.",
       from: 'en',
       to: 'de',
@@ -53,24 +63,15 @@ void main() async {
   //Die Anordnungen waren sehr kompliziert und schwierig.
 
   ///without source language (auto) and choosing a random instance:
-  libTransl = await lt.translate(
+  libTransl = await lt.translateSimply(
       "The dispositions were very complicated and difficult.",
       to: 'de',
       instanceMode: InstanceMode.Random);
   print(libTransl.translations.text);
   //Die Anordnungen waren sehr kompliziert und schwierig.
 
-  ///using Googletranslate:
-  ///short form to only get translated text as String, also shorter code:
-  textResult = await gt.tr("Er läuft schnell.", "en", 'de');
-
-  ///is the same as
-  textResult = await gt.tr("Er läuft schnell.");
-  print(textResult);
-  //He walks fast.
-
   ///long form to also get definitions and single word translations and switching to next instance, 4 retries if it fails (default 1)
-  var gtransl = await gt.translate(
+  var gtransl = await gt.translateSimply(
       "The dispositions were very complicated and difficult.",
       from: 'en',
       to: 'de',
@@ -83,14 +84,14 @@ void main() async {
   //Die Dispositionen waren sehr kompliziert und schwierig.
 
   ///without source language (auto):
-  gtransl = await gt.translate(
+  gtransl = await gt.translateSimply(
       "The dispositions were very complicated and difficult.",
       to: 'de');
   //Die Dispositionen waren sehr kompliziert und schwierig.
 
   ///get multiple word translations in target language from Google
   ///returns Map<String, dynamic>
-  gtransl = await gt.translate("exuberance", from: 'en', to: 'de');
+  gtransl = await gt.translateSimply("exuberance", from: 'en', to: 'de');
   print(gtransl.translations.rawTranslations);
   //{adjective: {dick: {frequency: 1/3, words: [thick, fat, large, big, heavy, stout]}, faustdick: {frequency: 1/3,...
 
@@ -109,11 +110,12 @@ void main() async {
 
   ///changing instance each time
   for (int i = 0; i < gt.getInstances.length; i++) {
-    gtransl = await gt.translate("Gewechselt",
+    gtransl = await gt.translateSimply("Gewechselt",
         from: "de", instanceMode: InstanceMode.Random);
     print(gt.getCurrentInstance);
     //translate.josias.dev
     // translate.namazso.eu
     // translate.riverside.rocks,...
   }
+
 }